+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this package are documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2022-01-23
+
+### Changed
+- Casing of Longstring has been changed to LongString.
+
+### Fixed
+- LongString not properly overriding ToString()
+
 ## [1.0.1] - 2022-01-22
 
 ### Added